Tidy up code for consistency + Fix Player Turn Issue + Fix Listeners
Changed the public variables accessed through the inspector to private and also fixes an issue where the first player would occasionally be able to play twice. Also fixes an issue where the buttons would subscribe to an event multiple times and not unsubscribe, causing the event to fire multiple times.